compose steering actuator. remove old node versions.
