+++ src/hardware/steering_actuator/CMakeLists.txt
@@ -5,38 +5,50 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic -Werror)
 endif()
 
-# find dependencies
 find_package(ament_cmake REQUIRED)
-# uncomment the following section in order to fill in
-# further dependencies manually.
 find_package(ackermann_msgs REQUIRED)
 find_package(driverless_msgs REQUIRED)
 find_package(std_msgs REQUIRED)
 find_package(rclcpp REQUIRED)
+find_package(rclcpp_components REQUIRED)
 
-include_directories(include)
-include_directories(${CMAKE_SOURCE_DIR}/../../hardware/CAN_Common/include/)
-include_directories(${CMAKE_SOURCE_DIR}/../../common/driverless_common/include)
+include_directories(
+  include
+  ${CMAKE_SOURCE_DIR}/../../hardware/CAN_Common/include/
+  ${CMAKE_SOURCE_DIR}/../../common/driverless_common/include
+)
 
 set (SOURCES
   ${CMAKE_SOURCE_DIR}/../../hardware/CAN_Common/src/canopen.cpp
   ${CMAKE_SOURCE_DIR}/../../hardware/CAN_Common/src/can_interface.cpp
 )
 
-add_executable(steering_actuator_node src/node_steering_actuator.cpp ${SOURCES})
-add_executable(steering_actuator_test_node src/node_steering_actuator_test.cpp ${SOURCES})
+# ROS 2 components
+add_library(steering_actuator_component SHARED src/component_steering_actuator.cpp ${SOURCES})
+ament_target_dependencies(steering_actuator_component
+  "rclcpp"
+  "rclcpp_components"
+  "std_msgs"
+  "driverless_msgs"
+  "ackermann_msgs"
+)
+rclcpp_components_register_nodes(steering_actuator_component "steering_actuator::SteeringActuator")
 
-ament_target_dependencies(steering_actuator_node rclcpp ackermann_msgs driverless_msgs std_msgs)
-ament_target_dependencies(steering_actuator_test_node rclcpp ackermann_msgs driverless_msgs std_msgs)
+# ROS 2 nodes
+add_executable(steering_actuator_node src/node_steering_actuator.cpp)
+target_link_libraries(steering_actuator_node steering_actuator_component)
+ament_target_dependencies(steering_actuator_node
+  "rclcpp"
+)
 
-target_include_directories(steering_actuator_node PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
-target_include_directories(steering_actuator_test_node PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
+install(TARGETS
+  steering_actuator_component
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin)
 
-  install(TARGETS steering_actuator_node steering_actuator_test_node
+install(TARGETS
+  steering_actuator_node
   DESTINATION lib/${PROJECT_NAME})
 
 # config folder
