+++ src/hardware/steering_actuator/launch/machine.launch.py
@@ -8,7 +8,7 @@ def generate_launch_description():
         [
             Node(
                 package="steering_actuator",
-                executable="steering_actuator_test_node",
+                executable="steering_actuator_node",
                 parameters=[
                     get_package_share_path("steering_actuator") / "config" / "steering.yaml",
                 ],
